requirements.txt
# LiveBetter API dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
psycopg2-binary==2.9.9
pandas>=2.2.0
python-dotenv==1.0.0
numpy>=1.26.0
requests==2.31.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
